task7_job_tracker/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlmodel==0.0.22
sqlalchemy==2.0.31
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
